Move service into data layer
