+++ interface/controller/curriculum_controller.go
@@ -2,8 +2,8 @@ package controller
 
 import (
 	"tat_gogogo/data/crawler/repository"
+	"tat_gogogo/data/crawler/service"
 	"tat_gogogo/domain/model"
-	"tat_gogogo/domain/service"
 	"tat_gogogo/usecase"
 )
