+++ interface/controller/login_controller.go
@@ -3,8 +3,8 @@ package controller
 import (
 	"log"
 	"tat_gogogo/data/crawler/repository"
+	"tat_gogogo/data/crawler/service"
 	"tat_gogogo/domain/model"
-	"tat_gogogo/domain/service"
 	"tat_gogogo/usecase"
 )
