+++ usecase/curriculum_usecase.go
@@ -1,9 +1,9 @@
 package usecase
 
 import (
+	"tat_gogogo/data/crawler/service"
 	"tat_gogogo/domain/model"
 	"tat_gogogo/domain/repository"
-	"tat_gogogo/domain/service"
 
 	"log"
 )
